Move PFA and Kubernetes version to operator-common to be able to use it in tests of operators
