+++ operator-common/src/main/java/io/strimzi/operator/KubernetesVersion.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.cluster.operator;
+package io.strimzi.operator;
 
 /**
  * Represents Kubernetes version which CO runs on
